Restructed project to be a little more organized
